refactor(DEV-10395): `<InvoiceStatusChip/>` to separate directory
