+++ packages/sdk-react/src/components/receivables/InvoiceStatusChip/index.ts
@@ -1,0 +1,4 @@
+export {
+  InvoiceStatusChip,
+  type InvoiceStatusChipProps,
+} from './InvoiceStatusChip';
